Moved exercise-page component into components folder
