+++ src/app/app.module.ts
@@ -7,7 +7,7 @@ import { AppComponent } from './app.component';
 import { HttpClientModule } from '@angular/common/http';
 import { RegistrationPageComponent } from './components/registration-page/registration-page.component';
 import { LoginPageComponent } from './components/login-page/login-page.component';
-import { ExercisePageComponent } from './exercise-page/exercise-page.component';
+import { ExercisePageComponent } from './components/exercise-page/exercise-page.component';
 
 
 @NgModule({
